feat: use OnboardingPage hook in the ProfileNameScreen
