+++ cardstack/src/components/OnboardingPage/index.ts
@@ -1,0 +1,2 @@
+export { default as OnboardingPage } from './OnboardingPage';
+export { default as useOnboardingPage } from './useOnboardingPage';
