+++ cardstack/src/components/index.ts
@@ -54,4 +54,4 @@ export * from './ValidationMessage';
 export * from './FloatingTag';
 export { default as SuffixedInput } from './Input/SuffixedInput/SuffixedInput';
 export * from './BackupStatus';
-export { default as OnboardingPage } from './OnboardingPage/OnboardingPage';
+export * from './OnboardingPage';
